+++ app/database/__init__.py
@@ -1,8 +1,9 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, Session
-from app.config import settings
+from sqlalchemy.ext.declarative import declarative_base
 
-engine = create_engine(settings.DATABASE_URL)
+Base = declarative_base()
+engine = create_engine("sqlite:///./app.db")
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 def get_db() -> Session:
