+++ app/ui/pages/__init__.py
@@ -1,10 +1,11 @@
 """UI pages initialization"""
-from .dashboard import render_page as dashboard
+from typing import Optional, List
 from .project_management import render_page as project_management
 from .csp_lob_management import render_page as csp_lob_management
+from .y_line_management import render_page as y_line_management
 from .competitor_management import render_page as competitor_management
 from .service_area_management import render_page as service_area_management
-from .y_line_management import render_page as y_line_management
+from .dashboard import render_page as dashboard
 
 __all__ = [
     'dashboard',
