+++ tests/integration/services/test_project_integration.py
@@ -1,0 +1,10 @@
+"""Integration tests for project service"""
+import pytest
+from sqlalchemy.orm import Session
+from app.models.project import Project, ProjectStatus
+from app.schemas.project import ProjectCreate
+from app.schemas.csp_lob import CSPLOBCreate
+from .base_service_test import BaseServiceTest
+
+class TestProjectIntegration(BaseServiceTest):
+    def test_complete_project_workflow(self):
