--- tests/test_basic.py
@@ -1,5 +1,0 @@
-def test_simple():
-    assert True
-
-def test_addition():
-    assert 1 + 1 == 2
