+++ tests/ui/test_pages.py
